+++ packages/eslint-plugin/src/index.ts
@@ -85,6 +85,9 @@ import preferOutputEmitterRef, {
 import preferOutputReadonly, {
   RULE_NAME as preferOutputReadonlyRuleName,
 } from './rules/prefer-output-readonly';
+import preferServiceDecorator, {
+  RULE_NAME as preferServiceDecoratorRuleName,
+} from './rules/prefer-service-decorator';
 import preferSignalModel, {
   RULE_NAME as preferSignalModelRuleName,
 } from './rules/prefer-signal-model';
@@ -142,6 +145,9 @@ import noImplicitTakeUntilDestroyed, {
 import computedMustReturn, {
   RULE_NAME as computedMustReturnRuleName,
 } from './rules/computed-must-return';
+import injectAtTop, {
+  RULE_NAME as injectAtTopRuleName,
+} from './rules/inject-at-top';
 
 export = {
   rules: {
@@ -154,6 +160,7 @@ export = {
     [contextualLifecycleRuleName]: contextualLifecycle,
     [directiveClassSuffixRuleName]: directiveClassSuffix,
     [directiveSelectorRuleName]: directiveSelector,
+    [injectAtTopRuleName]: injectAtTop,
     [noAsyncLifecycleMethodRuleName]: noAsyncLifecycleMethod,
     [noAttributeDecoratorRuleName]: noAttributeDecorator,
     [noDeveloperPreviewRuleName]: noDeveloperPreview,
@@ -180,6 +187,7 @@ export = {
       preferOnPushComponentChangeDetection,
     [preferOutputEmitterRefRuleName]: preferOutputEmitterRef,
     [preferOutputReadonlyRuleName]: preferOutputReadonly,
+    [preferServiceDecoratorRuleName]: preferServiceDecorator,
     [preferSignalModelRuleName]: preferSignalModel,
     [preferSignalsRuleName]: preferSignals,
     [preferStandaloneRuleName]: preferStandalone,